requirements/base.txt
Django==4.2
environs[django]==9.5.0
django-jazzmin==2.6.0
drf-yasg==1.21.5
django-phonenumber-field==7.0.2
phonenumbers==8.13.9
Pillow==9.5.0
psycopg2-binary==2.9.1
django-jazzmin==2.6.0
djangorestframework-simplejwt==5.2.2
django-admin-thumbnails==0.2.6
django-import_export == 3.2.0
django-modeltranslation==0.17.1
django-ckeditor==6.1.0
django-rosetta ==0.9.8
django-import_export == 3.2.0
<<<<<<< HEAD
django-recaptcha==3.0.0
django-filter==23.1
=======
captcha==0.5.0
django-simple-captcha==0.5.17
django-elasticsearch-dsl==7.3.0

>>>>>>> 16bb7480de51200b572b54bc1ef2415382730a2e
